add loading + server to edit teams
